requirements.txt
celery
Django
djangorestframework
gunicorn
psycopg2-binary
whitenoise
